+++ src/foreign/vigra_impex/imageinfo.cxx
@@ -99,7 +99,7 @@ void findImageSequence(const std::string &name_base,
 
 	// on Windows, both '/' and '\' are valid path separators
 	// note: std::basic_string.rfind() may return unsigned int, so exlicitely use std::max<int>()
-	size_t split = std::max<int>(name_base.rfind('/'), name_base.rfind('\\'));  
+	int split = std::max<int>(static_cast<int>(name_base.rfind('/')), static_cast<int>(name_base.rfind('\\')));  
     if(split == -1)
     {
         path = ".";
@@ -107,7 +107,7 @@ void findImageSequence(const std::string &name_base,
     }
     else
     {
-        for(size_t i=0; i<split; ++i)
+        for(int i=0; i<split; ++i)
         {
             if(name_base[i] == '/')
                 path += '\\';
